+++ jobbergate-api/CHANGELOG.rst
@@ -6,6 +6,7 @@ This file keeps track of all notable changes to jobbergate-api
 
 Unreleased
 ----------
+- Fix search when listing job-scripts
 
 4.0.0a1 -- 2023-08-15
 ---------------------